+++ docs/make.jl
@@ -16,6 +16,7 @@ makedocs(;
         assets=String[],
         collapselevel=1,
         sidebar_sitename=true,
+        size_threshold=1_000_000,
     ),
     plugins=[bib],
     pages=[
@@ -31,7 +32,9 @@ makedocs(;
                     "EFIE"=>"manual/examplesTH/efie.md",
                     "MFIE"=>"manual/examplesTH/mfie.md",
                 ],
-                "Time-Domain"=>Any["EFIE"=>"manual/examplesTD/tdefie.md"],
+                "Time-Domain"=>Any[
+                    "EFIE"=>"manual/examplesTD/tdefie.md",
+                ],
             ],
             "System of Equations and Bilinear Forms" => "manual/bilinear.md",
         ],
@@ -62,6 +65,13 @@ makedocs(;
                 "Monopole"=>"excitations/monopole.md",
                 "Linear Potential"=>"excitations/linearpotential.md",
             ],
+            "Composed Operator"=>Any[
+                "Composed Operator"=>"composedoperator/composedoperator.md"
+            ],
+            "Quasi Helmholtz Projectors"=>Any[
+                "Low Frequency Stable PMCHWT"=>"projectors/pmchwt_lf.md",
+                "Dense Grid Stable PMCHWT"=>"projectors/pmchwt_theta.md",
+            ],
         ],
         "Basis Functions" => Any[
             "Overview"=>"bases/overview.md",
@@ -73,7 +83,10 @@ makedocs(;
             ],
             "Temporal"=>Any[],
         ],
-        "Geometry Representations" => Any["Flat"=>"geometry/flat.md", "Curvilinear"=>Any[]],
+        "Geometry Representations" => Any[
+            "Flat"=>"geometry/flat.md",
+            "Curvilinear"=>Any[]
+        ],
         "____________________________________" => Any[],
         "Internals" => Any[
             "Overview"=>"internals/overview.md",
@@ -82,8 +95,8 @@ makedocs(;
             "Parametric Domain"=>Any[],
             "Multithreading"=>Any[],
         ],
-        "Contributing" => "contributing.md",
-        "References" => "references.md",
+        "Contributing" => Any["contributing.md"],
+        "References" => Any["references.md"],
         "API Reference" => "apiref.md",
     ],
 )