+++ test/runtests.jl
@@ -18,6 +18,7 @@ include("test_fourier.jl")
 include("test_specials.jl")
 
 include("test_basis.jl")
+include("test_lagrange.jl")
 include("test_directproduct.jl")
 include("test_raviartthomas.jl")
 include("test_rt.jl")
@@ -54,6 +55,10 @@ include("test_assemble_refinements.jl")
 include("test_dipole.jl")
 
 include("test_sauterschwabints1D.jl")
+include("test_hh2d_exec.jl")
+include("test_hh2d_mie.jl")
+include("test_hh2d_mie_higher_order.jl")
+include("test_hh2d_nearfield.jl")
 
 include("test_wiltonints.jl")
 include("test_sauterschwabints.jl")
@@ -82,15 +87,22 @@ include("test_variational.jl")
 
 include("test_handlers.jl")
 include("test_ncrossbdm.jl")
-include("test_curl_lagc0d1_lagc0d2.jl")
+#include("test_curl_lagc0d1_lagc0d2.jl")
 include("test_gridfunction.jl")
 
+include("test_itsolver.jl")
+
 include("test_hh_lsvie.jl")
+
+include("test_composed_basis.jl")
+include("test_composed_operator.jl")
+include("test_analytic_excitation.jl")
 include("test_vie.jl")
 include("test_evie_dvie.jl")
 
-@run_package_tests filter=ti->!(:example in ti.tags) verbose=true
+include("test_coloring.jl")
 
+@run_package_tests filter = ti -> !isempty(intersect([:example, :diagnostics], ti.tags)) verbose = true
 
 try
     Pkg.installed("BogaertInts10")